+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial release of the Viscalyx.GitHub PowerShell module
 - New command `Get-GitHubRelease` to retrieve releases from a GitHub repository
   with options to filter by latest, prerelease, and draft statuses
-- New command `Get-GitHubReleaseAssetMetadata` to retrieve metadata information
+- New command `Get-GitHubReleaseAsset` to retrieve metadata information
   about assets from GitHub repository releases
 - Private function `Convert-SecureStringAsPlainText` to safely handle secure
   string conversions for GitHub authentication tokens